添加微信、七牛基础功能、增加登录接口。未写单元测试，前端资源部署还需优化
