+++ lib/app.js
@@ -9,7 +9,7 @@ import bodyParser from 'body-parser';
 import db from './linkMongo';//链接数据库
 import index from './routes/index';
 import ajaxHandler from './routes/ajaxHandler';
-import wechat from './dbModule/wechat';//引入微信模块
+import wechat from './module/wechat';//引入微信模块
 
 let app = express();
 //拼接新路径，将public和view文件夹跟服务代码的文件夹(lib/dist)独立开
