--- lib/dbModule/wechat/jsApiList.js
@@ -1,31 +1,0 @@
-module.exports = {
-	'list':[
-		'onMenuShareTimeline',
-		'onMenuShareAppMessage',
-		'chooseImage',
-		'previewImage',
-		'uploadImage',
-		'downloadImage',
-		'startRecord',
-		'stopRecord',
-		'onVoiceRecordEnd',
-		'playVoice',
-		'pauseVoice',
-		'stopVoice',
-		'onVoicePlayEnd',
-		'uploadVoice',
-		'downloadVoice',
-		'translateVoice',
-		'getNetworkType',
-		'openLocation',
-		'getLocation',
-		'scanQRCode',
-		'hideOptionMenu',
-		'showOptionMenu',
-		'hideMenuItems',
-		'showMenuItems',
-		'hideAllNonBaseMenuItem',
-		'showAllNonBaseMenuItem',
-		'closeWindow'
-	]
-}
