--- lib/dbModule/wechat/menus.json
@@ -1,25 +1,0 @@
-{
-    "button":[
-        { 
-            "type":"click",
-            "name":"今日推荐",
-            "key":"today_recommend"
-        },
-        { 
-            "name":"小工具",
-            "sub_button":[{
-                "type": "scancode_waitmsg", 
-                "name": "扫一扫",
-                "key": "scancode"
-            },{
-                "type": "pic_sysphoto", 
-                "name": "系统拍照发图",
-                "key": "take_photo"
-            },{
-                "type": "location_select", 
-                "name": "发送位置",
-                "key": "send_location"
-            }]
-        }
-    ]
-}
