+++ lib/module/mongoose.js
