+++ lib/module/wechat/config.js
@@ -2,6 +2,6 @@ module.exports = {
 	token: '',     
 	appid:'',
 	appsecret:'',
-	encodingAESKey:'',   
+	encodingAESKey:'',      
 	checkSignature: false //可选，默认为true。由于微信公众平台接口调试工具在明文模式下不发送签名，所以如要使用该测试工具，请将其设置为false 
 }
