+++ lib/routes/index.js
@@ -18,4 +18,7 @@ router.get('/', function(req, res, next) {
 	// }
   res.render('index', { title: 'Express' });
 });
+router.get('/demo', function(req, res, next) {
+  res.render('demo', { title: 'demo' });
+});
 export default router;
