+++ package.json
@@ -20,7 +20,10 @@
     "multer": "^1.3.1",
     "node-uuid": "^1.4.8",
     "qiniu": "^7.2.1",
-    "serve-favicon": "~2.4.2"
+    "request": "^2.87.0",
+    "serve-favicon": "~2.4.2",
+    "wechat": "^2.1.0",
+    "wechat-api": "^1.35.1"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
